+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unpublished
 ### Added
 - add setting for include member birthdays into public ics export
+### Changed
+- use standard notification message for click on calendar link on dashboard
 
 ## v0.6.4 - 2024-02-11
 ### Added